Add most of the remaining controls, but they don't do anything yet
I added set_expand to Clear Coat and used it to get the lists filling
the window. Now set_top_level_margin_and_gap is called which really
helps the window not look bad. Also, now there is a name text box that
syncs back and forth with the selected job.